+++ test/test_ufl_wrapper.py
@@ -2,7 +2,7 @@ import pytest
 import ufl
 
 import basix
-from basix import ufl_wrapper
+import basix.ufl
 
 
 @pytest.mark.parametrize("inputs", [
@@ -12,7 +12,7 @@ from basix import ufl_wrapper
     (basix.ElementFamily.P, "triangle", 2),
 ])
 def test_create_element(inputs):
-    ufl_wrapper.create_element(*inputs)
+    basix.ufl.create_element(*inputs)
 
 
 @pytest.mark.parametrize("inputs", [
@@ -22,7 +22,7 @@ def test_create_element(inputs):
     (basix.ElementFamily.P, "triangle", 2),
 ])
 def test_create_vector_element(inputs):
-    ufl_wrapper.create_vector_element(*inputs)
+    basix.ufl.create_vector_element(*inputs)
 
 
 @pytest.mark.parametrize("inputs", [
@@ -32,7 +32,7 @@ def test_create_vector_element(inputs):
     (basix.ElementFamily.P, "triangle", 2),
 ])
 def test_create_tensor_element(inputs):
-    ufl_wrapper.create_tensor_element(*inputs)
+    basix.ufl.create_tensor_element(*inputs)
 
 
 @pytest.mark.parametrize("inputs", [
@@ -42,9 +42,9 @@ def test_create_tensor_element(inputs):
     (basix.ElementFamily.P, "triangle", 2),
 ])
 def test_tensor_element_hash(inputs):
-    e = ufl_wrapper.create_element(*inputs)
-    sym = ufl_wrapper.TensorElement(e, symmetric=True)
-    asym = ufl_wrapper.TensorElement(e, symmetric=None)
+    e = basix.ufl.create_element(*inputs)
+    sym = basix.ufl.TensorElement(e, symmetric=True)
+    asym = basix.ufl.TensorElement(e, symmetric=None)
     assert sym != asym
     assert hash(sym) != hash(asym)
 
@@ -59,7 +59,7 @@ def test_tensor_element_hash(inputs):
     ufl.EnrichedElement(ufl.VectorElement("Lagrange", "triangle", 1), ufl.VectorElement("Bubble", "triangle", 3)),
 ])
 def test_convert_ufl_element(e):
-    e2 = ufl_wrapper.convert_ufl_element(e)
+    e2 = basix.ufl.convert_ufl_element(e)
     # Check that element is hashable
     hash(e2)
 
@@ -70,40 +70,40 @@ def test_convert_ufl_element(e):
     ("Lagrange", "tetrahedron", 2, [])
 ])
 def test_converted_elements(celltype, family, degree, variants):
-    e1 = ufl_wrapper.create_element(celltype, family, degree, *variants)
+    e1 = basix.ufl.create_element(celltype, family, degree, *variants)
     e2 = ufl.FiniteElement(celltype, family, degree)
-    assert e1 == ufl_wrapper.convert_ufl_element(e1)
-    assert e1 == ufl_wrapper.convert_ufl_element(e2)
+    assert e1 == basix.ufl.convert_ufl_element(e1)
+    assert e1 == basix.ufl.convert_ufl_element(e2)
 
-    e1 = ufl_wrapper.create_vector_element(celltype, family, degree, *variants)
+    e1 = basix.ufl.create_vector_element(celltype, family, degree, *variants)
     e2 = ufl.VectorElement(celltype, family, degree)
-    assert e1 == ufl_wrapper.convert_ufl_element(e1)
-    assert e1 == ufl_wrapper.convert_ufl_element(e2)
+    assert e1 == basix.ufl.convert_ufl_element(e1)
+    assert e1 == basix.ufl.convert_ufl_element(e2)
 
 
 @pytest.mark.parametrize("elements", [
     [ufl.FiniteElement("Lagrange", "triangle", 1), ufl.FiniteElement("Bubble", "triangle", 3)],
-    [ufl.FiniteElement("Lagrange", "quadrilateral", 1), ufl_wrapper.create_element("Bubble", "quadrilateral", 2)],
+    [ufl.FiniteElement("Lagrange", "quadrilateral", 1), basix.ufl.create_element("Bubble", "quadrilateral", 2)],
     [ufl.VectorElement("Lagrange", "quadrilateral", 1),
-     ufl_wrapper.create_vector_element("Bubble", "quadrilateral", 2)],
+     basix.ufl.create_vector_element("Bubble", "quadrilateral", 2)],
     [ufl.TensorElement("Lagrange", "quadrilateral", 1),
-     ufl_wrapper.create_tensor_element("Bubble", "quadrilateral", 2)],
+     basix.ufl.create_tensor_element("Bubble", "quadrilateral", 2)],
 ])
 def test_enriched_element(elements):
-    e = ufl_wrapper._create_enriched_element([ufl_wrapper.convert_ufl_element(e) for e in elements])
+    e = basix.ufl._create_enriched_element([basix.ufl.convert_ufl_element(e) for e in elements])
     # Check that element is hashable
     hash(e)
 
 
 @pytest.mark.parametrize("e,space0,space1", [
-    (ufl_wrapper.create_element("Lagrange", basix.CellType.triangle, 2), "H1", basix.SobolevSpace.H1),
-    (ufl_wrapper.create_element("Discontinuous Lagrange", basix.CellType.triangle, 0),
+    (basix.ufl.create_element("Lagrange", basix.CellType.triangle, 2), "H1", basix.SobolevSpace.H1),
+    (basix.ufl.create_element("Discontinuous Lagrange", basix.CellType.triangle, 0),
      "L2", basix.SobolevSpace.L2),
-    (ufl_wrapper.MixedElement((ufl_wrapper.create_element("Lagrange", basix.CellType.triangle, 2),
-                               ufl_wrapper.create_element("Lagrange", basix.CellType.triangle, 2))),
+    (basix.ufl.MixedElement((basix.ufl.create_element("Lagrange", basix.CellType.triangle, 2),
+                               basix.ufl.create_element("Lagrange", basix.CellType.triangle, 2))),
      "H1", basix.SobolevSpace.H1),
-    (ufl_wrapper.MixedElement((ufl_wrapper.create_element("Discontinuous Lagrange", basix.CellType.triangle, 2),
-                               ufl_wrapper.create_element("Lagrange", basix.CellType.triangle, 2))),
+    (basix.ufl.MixedElement((basix.ufl.create_element("Discontinuous Lagrange", basix.CellType.triangle, 2),
+                               basix.ufl.create_element("Lagrange", basix.CellType.triangle, 2))),
      "L2", basix.SobolevSpace.L2),
 ])
 def test_sobolev_space(e, space0, space1):
